Consolidated MFT Transfer nodes to a single node
