--- nodes/mft-transfer-retrieve.html
@@ -1,106 +1,0 @@
-<!--
-  Copyright 2022 IBM Corp.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-
-<script type="text/html" data-template-name="mft-transfer-retrieve">
-  <div class="form-row">
-    <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-    <input type="text" id="node-input-name" placeholder="Name">
-  </div>
-  
-  <div class="form-row">
-    <label for="node-input-user"><i class="fa fa-folder-close"></i> User</label>
-    <select id="node-input-user" style="width:68%;">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-server"><i class="fa fa-folder-close"></i> Server</label>
-    <select id="node-input-server" style="width:68%">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-apiv"><i class="fa fa-folder-close"></i> API Version</label>
-    <input type="text" id="node-input-apiv">
-  </div>
-
-  <div class="form-tips">
-    <span><b>Note: </b>This is an admin node, so please ensure the user selected has admin authorisation</span>
-    <br>
-    <span>This node allows the following parameters via an inject node:</span>
-    <ul>
-      <li><b>msg.id</b> : The id of the transfer to retieve details for.</li>
-    </ul>
-  </div>
-
-</script>
-
-<script type="text/html" data-help-name="mft-transfer-retrieve">
-  <p>This node makes use of the MQ REST API to retrieve details of transfers from the Queue Manager. </p>
-
-  <p>Specify the expected content type in the configuration. The node will check that
-  the contents of <code>msg.payload</code> matches the specified type.</p>
-</script>
-
-<style type="text/css">
-    .node_label_white {
-        fill: white;
-    }
-
-    #palette_node_mqrest-out>div.palette_label {
-        color: white;
-    }
-</style>
-
-<script type="text/javascript">
-    (function () {
-        RED.nodes.registerType('mft-transfer-retrieve', {
-            category: 'IBM MQ REST',
-            defaults: {
-                name: { value: '' },
-                user: { value: "", type: "user-config" },
-                server: { value: "", type: "server-config"},
-                apiv: { value: 'v2'}
-            },
-            credentials: {
-            },
-            color: '#C7E9C0',
-            inputs: 1,
-            outputs: 1,
-            paletteLabel: 'MFT Transfer (rtv)',
-            icon: 'feed.png',
-            label: function () {
-                return this.name || 'MFT Transfer (rtv)';
-            },
-            labelStyle: function () {
-                return this.name ? 'node_label_italic node_label_black' : 'node_label_black';
-            },
-            oneditprepare: function () {
-              $("#node-input-apiv").typedInput({
-                types: [
-                  {
-                    value: "version",
-                    options: [
-                      { value: "v1", label: "1" },
-                      { value: "v2", label: "2" }
-                    ]
-                  }
-                ]
-              })
-            }
-        });
-    })();
-</script>
